LIBITD-2704. Refactored common ProspectController code into concern
Refactored duplicate private methods related to ProspectsController
params handling out of the "ProspectsController" and
"Admin::ProspectsController" into a "ProspectParameterHandling" concern.

https://umd-dit.atlassian.net/browse/LIBITD-2704
